use an explicit cast to avoid compiler warnings
git-svn-id: https://svn.r-project.org/R/trunk@80844 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/nmath/bd0.c
@@ -248,7 +248,7 @@ void attribute_hidden ebd0(double x, double M, double *yh, double *yl)
 	// prevent later overflow
 	if (M_LN2 * ((double) -e)  > 1. + DBL_MAX / x) { *yh = ML_POSINF; return; }
 
-	int i = floor ((r - 0.5) * (2 * N) + 0.5);
+	int i = (int) floor ((r - 0.5) * (2 * N) + 0.5);
 	// now,  0 <= i <= N
 	double f = floor (S / (0.5 + i / (2.0 * N)) + 0.5);
 	double fg = ldexp (f, -(e + Sb)); // ldexp(f, E) := f * 2^E
